feature(#5283): add action log watcher retries + add fatal error in API workers + api refactoring and fixes

+++ community/go-engines-community/lib/canopsis/pbehavior/event_computer.go
@@ -209,7 +209,7 @@ func (c *eventComputer) computeByActiveType(
 
 func (c *eventComputer) sortExdates(exdates []Exdate) error {
 	var err error
-	getPriorityByID := func(id string) int {
+	getPriorityByID := func(id string) int64 {
 		t, ok := c.typesByID[id]
 		if !ok {
 			err = fmt.Errorf("unknown type %v", id)
